@@ -4,134 +4,164 @@ using CalculatorModel;
 using Microsoft.Extensions.Logging;
 using MVVM;
 using Patterns.Implementation;
+using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Windows.Input;
 
-namespace Calculator.MVVM.ViewModels.Standard
+namespace Calculator.MVVM.ViewModels.Standard;
+
+public class BottomViewModelLoadable<TItem> : Loadable<TItem> where TItem : Operation
 {
-    public class BottomViewModelLoadable<TItem> : Loadable<TItem> where TItem : Operation
+    private bool _isLblVisible;
+    public bool IsLblVisible
     {
-        private bool _isLblVisible;
-        public bool IsLblVisible
-        {
-            get => _isLblVisible;
-            set => UpdateObservable(ref _isLblVisible, value);
-        }
-        private string _nbr;
-        public string Nbr
-        {
-            get => _nbr;
-            set => UpdateObservable(ref _nbr, value);   
-        }
-        public bool IsPresent(string inputText)
-        {
-            Operation op = Items.FirstOrDefault(o => o.OpValue.Equals(inputText));
-            return op == null;
-        }
-        
-        public override void SetItems(IEnumerable<TItem> items)
-        {
-            base.SetItems(items);
-            IsLblVisible = !IsEmpty;
-        }
-        public override void SelectedItemCallBack(TItem selectedItem)
-        {
-            if(SelectedItem != null)
-            {
-                setValue(selectedItem);
-                ServiceLocator.StandardCalculatorViewModel.IsBottomDrawerOpen = false;
-            }
-        }
-        private void setValue(TItem v)
-        {
-            ServiceLocator.InputResultViewModel.InputText = v.OpValue;
-            ServiceLocator.InputResultViewModel.OutputText = v.OpResult;
-        }
+        get => _isLblVisible;
+        set => UpdateObservable(ref _isLblVisible, value);
+    }
+    private string _nbr;
+    public string Nbr
+    {
+        get => _nbr;
+        set => UpdateObservable(ref _nbr, value);
+    }
+    public bool IsPresent(string inputText)
+    {
+        Operation op = Items.FirstOrDefault(o => o.OpValue.Equals(inputText));
+        return op == null;
     }
 
-    public class OperationListViewModel: BottomViewModelLoadable<Operation>
+    protected override void SetItems(IEnumerable<TItem> items)
     {
-        private readonly ISettingsManager _settings;
-        private readonly int _threshold;
-        public OperationListViewModel(ISettingsManager settings)
+        base.SetItems(items);
+        IsLblVisible = !IsEmpty;
+    }
+    protected void SelectedItemCallBack(TItem selectedItem)
+    {
+        if (SelectedItem != null)
         {
-            _settings = settings;
-            _threshold = (int)_settings.GetParameter("CountThreshold");
-            SetSortDescription(new SortDescription("Id", ListSortDirection.Descending));
-            Init();
+            setValue(selectedItem);
+            ServiceLocator.StandardCalculatorViewModel.IsBottomDrawerOpen = false;
         }
-        private async void Init()
-        {
-            await Task.Run(async () =>
-            {
-                using CalculatorRepository repo = new CalculatorRepository();
-                await LoadItems(repo.GetAll());
-                IsLblVisible = !IsEmpty;
-            });
-        }
-        #region Override Methods
-        public override void DeleteAllItems()
-        {
-            using var repo = new CalculatorRepository();
-            repo.DeleteAllAsync();
-            base.DeleteAllItems();
-        }
-        protected override void NumberOfItemsChanged(int count)
-        {
-            if (count >= _threshold)
-            {
-                Nbr = $"{_threshold}+";
-            }
-            else
-            {
-                Nbr = $"{count}";
-            }
-        }
-        #endregion
+    }
+    private void setValue(TItem v)
+    {
+        ServiceLocator.InputResultViewModel.SetOpValue(v.OpValue, v.OpResult);
+    }
+}
+
+public class OperationListViewModel : BottomViewModelLoadable<Operation>
+{
+    private readonly ISettingsManager _settings;
+    private readonly int _threshold;
+
+
+    public OperationListViewModel(ISettingsManager settings)
+    {
+        _settings = settings;
+        OnSelectedIem = SelectedItemCallBack;
+        OnNumberOfItemsChanged = NumberOfItemsChanged;
+        _threshold = (int)_settings.GetParameter("CountThreshold");
+        SetSortDescription(new SortDescription("Id", ListSortDirection.Descending));
     }
 
-    public class BottomViewModel: BaseViewModel
+    public ObservableCollection<Operation> GetAllItems()
     {
-        #region Private properties
-        
-        private readonly ISettingsManager _settings;
-        private readonly ILogger<BottomViewModel> _logger;
+        return GetItems();
+    }
 
-        private OperationListViewModel _operationVM;
-        public OperationListViewModel OperationVM
+    public void DeleteAllItemsAsyn()
+    {
+        DeleteAllItems();
+    }
+    public void Add(string inputText, string ouputText)
+    {
+        using var repo = new CalculatorRepository();
+        var op = repo.Save(new(inputText, ouputText));
+        AddItem(op);
+    }
+    public async Task LoadAsync()
+    {
+        if(!IsEmpty)
         {
-            get => _operationVM;
-            set => UpdateObservable(ref _operationVM, value);
+            return;
         }
-
-        #endregion
-
-        #region Commands
-        public ICommand DeleteAllCommand { get; private set; }
-        #endregion
-
-        #region Constructor
-        public BottomViewModel(ILogger<BottomViewModel> logger, ISettingsManager settings)
+        using CalculatorRepository repo = new CalculatorRepository();
+        await LoadItems(repo.GetAll());
+        IsLblVisible = !IsEmpty;
+    }
+    #region Override Methods
+    protected override void DeleteAllItems()
+    {
+        using var repo = new CalculatorRepository();
+        repo.DeleteAllAsync();
+        base.DeleteAllItems();
+    }
+    protected void NumberOfItemsChanged(int count)
+    {
+        if (count >= _threshold)
         {
-            _logger = logger;
-            _settings = settings;
-            OperationVM = ServiceLocator.OperationListViewModel;
-            _logger.LogInformation("BottomViewModel started.....");
-            DeleteAllCommand = new DelegateCommand(OnDeleteAll);
+            Nbr = $"{_threshold}+";
         }
-        #endregion
-
-        #region Private methods
-        
-        #endregion
-
-        #region Handlers
-        private void OnDeleteAll(object parameter)
+        else
         {
-            OperationVM.DeleteAllItems();
+            Nbr = $"{count}";
         }
-        #endregion
+    }
+    #endregion
+}
 
-        
+public class BottomViewModel : BaseViewModel
+{
+    #region Private properties
+    
+    private readonly ISettingsManager _settings;
+    private readonly ILogger<BottomViewModel> _logger;
+    private OperationListViewModel _operationVM;
+
+
+    public ICommand LoadedCommand { get; private set; }
+    public OperationListViewModel OperationVM
+    {
+        get => _operationVM;
+        set => UpdateObservable(ref _operationVM, value);
+    }
+
+    #endregion
+
+    #region Commands
+    public ICommand DeleteAllCommand { get; private set; }
+    #endregion
+
+    #region Constructor
+    public BottomViewModel(ILogger<BottomViewModel> logger, ISettingsManager settings)
+    {
+        _logger = logger;
+        _settings = settings;
+        OperationVM = ServiceLocator.OperationListViewModel;
+        _logger.LogInformation("BottomViewModel started.....");
+        DeleteAllCommand = new DelegateCommand(OnDeleteAll);
+        LoadedCommand = new DelegateCommand(OnLoaded);
+    }
+    #endregion
+    private async void OnLoaded(object parameter)
+    {
+        await OperationVM.LoadAsync();
+    }
+
+    #region Private methods
+
+    #endregion
+
+    #region Handlers
+    private void OnDeleteAll(object parameter)
+    {
+        OperationVM.DeleteAllItemsAsyn();
+    }
+    #endregion
+
+    public void AddOperation(string inputText, string ouputText)
+    {
+        OperationVM.Add(inputText, ouputText);
+        OnPropertyChanged(nameof(OperationVM));
     }
 }
